requirements.txt
streamlit>=1.28
streamlit-authenticator>=0.3.0
numpy>=1.24
pandas>=2.0
scipy>=1.11
plotly>=5.18
openpyxl>=3.1
xlsxwriter>=3.1
pyyaml>=6.0
